+++ NEWS.md
@@ -1,10 +1,10 @@
 # daedalus 0.2.16
 
-Patch version to allow passing a list of infection objects to [daedalus()]. This helps model multiple different pandemics on the same population, or to model uncertainty in outcomes due to uncertainty in infection parameters for a single pandemic pathogen.
+Patch version to allow passing a list of infection objects to the Daedalus model using the new function `daedalus_multi_infection()`. This helps model multiple different pandemics on the same population, or to model uncertainty in outcomes due to uncertainty in infection parameters for a single pandemic pathogen.`daedalus_multi_infection()` returns a list of output objects.
 
-[daedalus()] returns a list of output objects when a list of infection objects is passed.
+- Added a vignette showing how to use `daedalus_multi_infection()`;
 
-- Added a vignette showing how to use infection list functionality.
+- Added functions to more cleanly handle inputs to `daedalus()` and `daedalus_multi_infection()`;
 
 - Some functions used in v0.1.0 have been removed.